move counter to include
