+++ .gitmodules
@@ -1,10 +1,12 @@
 [submodule "galois"]
 	path = galois
 	url = https://github.com/utcs-scea/Galois.git
-  branch = dgraph
+  	branch = dgraph
+
 [submodule "CombBLAS"]
 	path = CombBLAS
 	url = https://github.com/PASSIONLab/CombBLAS.git
+
 [submodule "pytorch"]
 	path = pytorch
 	url = https://github.com/utcs-scea/pytorch.git
